Add FBSDKGamingServicesKitTests to the BuildAllKits scheme
Summary:
The BuildAllKits test didn't include FBSDKGamingServicesKitTests. (It had FBSDKCoreKitTests, FBSDKLoginKitTests and FBSDKShareKitTests).

The FBSDKGamingServicesKitTests were still run via `buck test :`, however they would not run in Xcode when using Command+U to run tests for the scheme.

The compilation error I was initally getting was due to `carthage bootstrap` not completing, which is a known issue that is currently open https://github.com/Carthage/Carthage/issues/3019

The following workaround of using a custom `carthage.sh` file worked for me for that issue
https://github.com/Carthage/Carthage/issues/3019#issuecomment-665136323

Reviewed By: joesus

Differential Revision: D24778344

fbshipit-source-id: b105a9441439caf5f27c8d7276d2bed119de64c8