Convert FBSDKCoreKitTests/FBSDKAppLinkUtilityTests.m to Swift
Summary:
Converted FBSDKAppLinkUtilityTests.m to Swift

The FBSDKCoreKitTests-Bridging-Header.h file already existed but wasn't included in the project. Added it to the "Supporting Files" group and added a build configuration reference to it in FBSDKCoreKitTests.xcconfig

Also added a test for isMatchURLScheme to ensure that NSBundle.mainBundle could be mocked in Swift.

Reviewed By: joesus

Differential Revision: D24766482

fbshipit-source-id: dacc5eab7e3603d88a8863bca5411ac73e168939