FBSDKBridgeAPI Unit Tests 1/n
Summary: The main point of interest here is probably that I created a separate header to declare some types that are only declared in `FBSDKBridgeAPI.m`. This seemed preferable to creating an internal header in the main project since it's only needed for tests. Normally would put these in the test file itself but it was becoming crowded.

Reviewed By: jawwad

Differential Revision: D24875594

fbshipit-source-id: 1ad072942e0487e31ceca56df6bd47feaf59334c