Proof of concept for having a Swift extension on an existing Objective-C test class
Summary:
Allowing for Swift extenions on tests will allow us to partially convert existing test classes to Swift and start writing new tests in Swift.

The only caveat is that for the original test to be visible to the Swift extentsion of the test, its header needs to be part of the bridging header. Since the interface declaration for an XCTest isn't in a separate .h file but included within the .m file for the test there isn't a header file that can be included. However the actual interface declaration can be copied to the briding header which feels like a cleaner solution than to create a new .h file for inclusion in both the header and the original test.

Reviewed By: joesus

Differential Revision: D24867053

fbshipit-source-id: 51ba16e45e7d492573e5985f7eb3a82effc30574