Convert FBSDKAppEventsConfigurationTests.m to Swift
Summary: One minor modification is that the Objective-C version of the test was setting the config to nil in the tearDown method, however this wasn't strictly necessary as it was alrady being reset in the setUp method. Also setting it to nil prevented this variable from being non-nullable so removed the tearDown method.

Reviewed By: joesus

Differential Revision: D24943362

fbshipit-source-id: 4293d179d4e44d40e4363038a683ac6ea88f0467