Convert FBSDKAppEventsConfigurationManagerTests.m to Swift
Summary: This file is one that needs access to a private method in FBSDKAppEventsConfigurationManager. To expose the private method to Swift I added an entry for the category in the bridging header. We could decide to move these into a separate file if needed but I think it feels ok to have them directly in here for now. (Easier to search for bridging header than "what's the name of that new swift header I had created" for Objective-C categories that Swift needs")

Reviewed By: joesus

Differential Revision: D25031348

fbshipit-source-id: 236f95968ce17491da08bcda9e884248e4e7916c