Update the Carthage version of OCMock to match the internal version
Summary:
OCMock was upgraded internally for buck to 3.6.1 with D21750024. However the version referenced the in Cartfile is 3.5.1. Updated the version in the Cartfile to match the internal version so that we don't run into weird issues when building via BUCK vs Xcode.

This also takes care of the following error that was caused due to using an older version:
```
The iOS Simulator deployment target 'IPHONEOS_DEPLOYMENT_TARGET' is set to 8.0, but the range of supported deployment target versions is 9.0 to 14.2.99.
```

Also updated OCMVerify to OCMVerifyAll to fix 2 "Expression result unused" warnings
https://stackoverflow.com/a/63926392/340508

Reviewed By: joesus

Differential Revision: D25032351

fbshipit-source-id: 01fa64300359b8b3728274daeb25806af35b2fe5