FBSDKBridgeAPI Unit Tests 7/n
Summary:
This one is a bit different. Currently there's no way to mock the way we load dynamic frameworks. So we can only test one condition branch for every dynamic framework.

This fixes that by introducing a new protocol, `FBSDKDynamicFrameworkResolving` that wraps the c functions in the `FBSDKDynamicFrameworkLoader` class.

For call sites we introduce a private method override to pass in a `id<FBSDKDynamicFrameworkResolving>` object. The existing implementation passes in the shared instance of `FBSDKDynamicFrameworkLoader` and the new override can be used for testing.

This is a common practice for protocol-based unit testing in Swift and will make the eventual transition easier.

Reviewed By: jawwad

Differential Revision: D25034818

fbshipit-source-id: 10e5f2107712c0b904ddefae6647b50cbf1afe87