ID Token signature verification
Summary:
According to the openid connect spec (https://openid.net/specs/openid-connect-core-1_0.html#IDTokenValidation), the client must validate the ID token. One of the required verification steps is to validate the signature using the specified algorithm, in our case this will be RS256.

The diff add signature verification for ID token. An ID token can only be created if it has a valid signature.

Reviewed By: ppansy

Differential Revision: D24716889

fbshipit-source-id: 203116fdc8564453214a642b27abaf1f297fd280