Tweak Swift Interface
Summary:
The Swift typeahead should only show the method that works well with Swift.

Previously the signature:
`logIn(from:configuration:completion:)` was available for Swift but referred to the ObjC method which has a `(response, error)` completion type. This is not Swifty.

We also provided them with the signature:
`logIn(viewController:configuration:completion:)` which we want them to use because it has a `(result)` completion type which is very Swifty.

This removes the first signature from typeahead and prevents it from being called without explicit `__`. This is a better experience.

Reviewed By: ppansy

Differential Revision: D25284727

fbshipit-source-id: 1524f5f14d8d6ec42839c00e933ee9f72fc527d9