Login button Updates
Summary:
Integrates profile changes with login button. `FBSDKLoginButton` was solely depending on `FBSDKAccessToken` as a proxy for authentication but now it also looks at `FBSDKAuthenticationToken` since this with OICD this is a valid means of determining whether a login attempt was successful.

Also adds support for responding to profile notifications.

This creates an obvious issue because now we have a race condition between `Profile` and `AccessToken` notifications. This means we might wind up with an extra fetch if we handle the profile notification (which doesn't trigger a fetch) and then handle get the access token notification which does trigger a fetch.

However, sometimes this is what we want since the information fetched by handling the access token might be more current than that of the current profile (which might be served from an old cache).

So to solve this we could do a number of things:

a) Write a bunch of delayed dispatches and locks to wait until we've received both token and then handle use the right one using the datestamps on each.
b) Write some intermediate object that handles all of this outside of the button and dispatches a single notification to the button.
c) Hold onto a potential fetch task so we can cancel if it we get a profile notification with a recently refreshed profile.

I'm fairly sure that each of these will also cause their own edge cases so it's probably better to occasionally perform a needless fetch in order to keep everything easier to reason about.

Reviewed By: ppansy

Differential Revision: D25292523

fbshipit-source-id: 198725a66c2b15c1e77cc34dbfa428ab709d06f0