Add post-login heartbeat event to FB Login
Summary:
Currently our last client-side event is passing access token back to the app. If something suddenly breaks after this, we have no visibility.

Similar to D24403530, adding a new event that will be logged 5 seconds after passing token back to the app. If app breaks after login, this event won't be logged and we'll see a sudden drop.

Differential Revision: D25172326

fbshipit-source-id: af1a2de2a29b0e84d6815bf7539ee62078f69235