fix ProfilePictureView bug
Summary:
Fixing bug that ProfilePictureView is not getting cleared when a limited login session is invalidated by logout endpoint.

The bug is caused by two notifications trying to update profile picture at the same time:
1) On app launched, a saved Profile instance is read from cache, which triggers a notification
2) ProfilePictureView makes an async network call to fetch the image from Profile
3) "Not Authorized" status returned from logout endpoint, Profile is set to nil, which triggers another notification.
4) ProfilePictureView clears the image after seeing Profile is nil
5) image fetch from step 2) completed. ProfilePictureView sets the image using the fecthed data

The fix is to make sure `lastState` is set when the view will use an image, and is `nil` when the view will be cleared (use a placeholder). `_updateImageWithData:state` should return directly if the given state doesn't match `_lastState`. In the above example, `_lastState` will be set to nil in step 4), which will prevent the view from overwriting image in step 5)

Reviewed By: joesus, regresscheck

Differential Revision: D25650327

fbshipit-source-id: 59098376c93ea0f7e3cfc1355a2f8154e11d213a