Fix FBSDKLoginButton nonce property attributes (#1616)
Summary:
Thanks for proposing a pull request!

To help us review the request, please complete the following:

- [x] sign [contributor license agreement](https://developers.facebook.com/opensource/cla)
- [x] I've ensured that all existing tests pass and added tests (when/where necessary)
- [x] I've updated the documentation (when/where necessary) and [Changelog](CHANGELOG.md) (when/where necessary)
- [x] I've added the proper label to this pull request (e.g. `bug` for bug fixes)

## Pull Request Details

The attributes for the `nonce` property of the `FBSDKLoginButton` are incorrect and lead to build errors.

In scenarios where the [`objc-property-assign-on-object-type`](https://clang.llvm.org/docs/DiagnosticsReference.html#wobjc-property-assign-on-object-type) flag is enabled and warnings are treated as errors, the build fails with: `FBSDKLoginKit.framework/Headers/FBSDKLoginButton.h:109:1: 'assign' property of object type may become a dangling reference; consider using 'unsafe_unretained'`

Pull Request resolved: https://github.com/facebook/facebook-ios-sdk/pull/1616

Reviewed By: jingping2015

Differential Revision: D26002203

Pulled By: ppansy

fbshipit-source-id: 5101f92f8f79b069f63b022129f70ad7becdb4e3