Unit Test Hardening RestrictiveDataFilterTests
Summary:
The app events initializer will fetch a server configuration if the SDK is initialized. Additionally we don't disable the flush timer on the mock so it will fire if a test is slow running.

This fixes those issues for this test class. May need to revisit this solution as more edge cases come up during test refactoring but for now it works.

Reviewed By: Mxiim

Differential Revision: D26149574

fbshipit-source-id: f566aedefbc1655540e54a4db4d59f9ffc2085da