Fix Swift Package Manager error for Xcode 12.5 (#1635)
Summary:
Pull Request resolved: https://github.com/facebook/facebook-ios-sdk/pull/1635

12.5 Beta errors on loading a package with nested 'include' directories. At one point we needed this in order to access internal headers from CoreKit.

Checked this fix with Xcode 12.5_Beta, Xcode 11.2, and Xcode 11.7 so either that workaround was never needed or it was needed before 11.2 and was fixed.

Given the age of the software and the likelihood of Apple dropping support for Xcode 11 in the near future, there is little incentive to come up with a workaround for supporting below 11.2.

Reviewed By: Mxiim

Differential Revision: D26227309

fbshipit-source-id: e9f15599152c413891b03632d20479b948791642