AtePublisher Refactor (#1646)
Summary:
Pull Request resolved: https://github.com/facebook/facebook-ios-sdk/pull/1646

Abstracts the AtePublisher from AppEvents.

Reasons:
AppEvents is nearly a thousand lines long. The publishing of ATE is a unique concern that is related but has no direct dependencies on AppEvents. This simplifies the AppEvents tests because all we care about from the AppEvents side is

a) it has a default publisher of the expected type
b) it invokes that publisher as part of its publishATE method

This allows us to test the publisher in isolation by using an internal initializer that takes the required dependencies which include:

* NSUserDefaults for persisting the last publish date
* NSDate for finding out when the last publish date was
* FBSDKSettings for advertiser tracking status
* FBSDKGraphRequest for creating and starting the graph request
* FBSDKInternalUtility for getting the operating system version

This diff does the following:
1. Moves the body of the publishATE method to be a distinct type (this diff)
1. Tests that the type is used correctly by AppEvents (this diff)
1. Ports the existing tests for the new type (this diff)
1. Adds protocols for the dependencies of the new type so that it can be tested without swizzling. (future diff)
1. Uses fakes to test the new type (future diff)

Reviewed By: Mxiim

Differential Revision: D26341851

fbshipit-source-id: 55a68bd7b63f348990066f28b07f44052c1b4b42