Cleanup FBSDKSKAdNetworkConversionConfigurationTests
Summary:
Removes unused methods.
Precludes creation of graph requests for partial mocks of AppEvents.

This is the nuclear option but seems to work for now to keep the tests from creating and starting graph requests.

Ultimately the plan is to introduce an injectable GraphRequestProvider that can be stubbed out so we can do more robust testing around the created graph requests.

Reviewed By: jingping2015

Differential Revision: D26346061

fbshipit-source-id: 54cf5ecca28db876a7d37d26e1e3c38d148b3c31