Fix name conflict for URLSessionTask
Summary:
I'm on the fence about whether or not this is a breaking change. Technically it's a in a public header but it is also part of CoreKitBasics, a dependency of CoreKit and not a standalone library.

The question is, will this break existing implementations. I see how this is technically possible but also extremely unlikely. You would need to be trying to explicitly create a `FBSDKURLSessionTask` which is just a lightweight proxy wrapper around `NSURLSessionTask`.

It's also not listed in the public documentation so is by nature 'use at your own risk'.

Reviewed By: jingping2015

Differential Revision: D26429413

fbshipit-source-id: 472365cf3259f6c32d0b2b408fab63da4c9f74a3