Abstracting Networker from GraphRequestConnection 1/n
Summary:
Introduces two new internal protocols to handle resolving the `FBSDKGraphRequestConnection` dependency on `FBSDKURLSession`.

`FBSDKURLSessionProxying` - describes the `FBSDKURLSession` which is essentially just a proxy to `NSURLSession`'s behaviors.
`FBSDKURLSessionProxyProviding` - describes anything that can provide a `FBSDKURLSessionProxying`

Additionally introduces a new internal type `FBSDKURLSessionProxyProvider` which essentially just wraps the creation of `FBSDKURLSessions`'s and returns them as `FBSDKURLSessionProxying`'s.

`FBSDKGraphRequestConnection` has a default `FBSDKURLSessionProxyProvider`. The sessions used by `FBSDKGraphRequestConnection` are derived from this object.

This is a step towards true dependency injection since the `FBSDKGraphRequestConnection` now is asking up the chain for the session proxy to use instead of being provided with one directly.

The followup to this diff shows the real power of this, where we can test complex scenarios without relying on runtime swizzling.

Feedback welcome.

Reviewed By: dreamolight

Differential Revision: D26380778

fbshipit-source-id: 554196bf1cdd341815c306c47949abeca4840430