Deprecate FBSDKGraphErrorRecoveryProcessor.delegate
Summary:
With the previous work on FBSDKErrorRecoveryAttempting in place, we no longer need to keep a delegate property on FBSDKGraphErrorRecoveryProcessor, so it is now deprecated. There were several problems with the existing approach:

- `delegate` was defined as a `strong` property
- historically, some code paths never nilled out the property, creating retain cycles
- if multiple calls to `processError:` (with different `delegate` parameters) were made on the same instance of FBSDKGraphErrorRecoveryProcessor, problems would arise by tracking only one delegate instance over asynchronous recovery methods

Since FBSDKErrorRecoveryAttempting now takes a completion handler, we can pass the incoming delegate along to those methods, avoiding altogether the need to store (weakly or strongly) the delegate in the first place.

In addition, the `didPresentErrorWithRecovery:contextInfo:` declaration was erroneously public (and no longer used by FBSDK), so it is also deprecated.

Finally, in the case of a `FBSDKGraphRequestErrorOther` error, we weren't storing the error (into `self->_error`) for the UIAlertAction action to later use (similar to the `FBSDKGraphRequestErrorRecoverable` case). That is fixed.

Reviewed By: joesus

Differential Revision: D26491372

fbshipit-source-id: 5eaf8ac62dd3a078041784b2fa543cca59a36eb6