Adding GraphRequestConnectionFactory to GraphRequest
Summary:
Currently `GraphRequest` has a convenience method for creating a `GraphRequestConnection` adding itself to that connection, and starting the connection. Since there are dependencies on `GraphRequestConnection`, we need a way to pass these dependencies from the owner of the connection. In this case, the `GraphRequest`.

We've added a few new types to help satisfy this requirement.

Protocol `GraphRequestConnecting` to describe anything that behaves like a `GraphRequestConnection`
Protocol `GraphRequestConnectionProviding` to describe anything that returns a `GraphRequestConnecting`
Class `FBSDKGraphRequestConnectionFactory` that conforms to `GraphRequestConnectionProviding` and returns concrete instances of `GraphRequestConnection`.

So in our tests we can provide a factory that returns fake graph request connections that can be started, stopped, validated, etc...
And in our production code we can use the default connection factory to return the concrete instances we expect.

We'll continue to use this pattern to surface and push dependencies up the stack until they can be concentrated at a single injection point.

Additionally it was necessary to perform a small refactor on LoginKit to get the semantics right. (It was treating a graph request connection as a graph request connection **provider** instead of having separate protocols for differentiating a connection from something that provides a connection)

Reviewed By: ppansy

Differential Revision: D26471361

fbshipit-source-id: f93f1b6252822e5ca89dbec78251786dd23bc232