Backfill GraphRequestConnection Unit Tests 2/n
Summary:
This diff tries to solve the problem of 'how do we reduce our dependence on stateful static types'.

One strategy is to have a protocol define an instance method that mirrors a class method. Then have a new object wrap the existing singleton, conform to that protocol and the dependent type takes an instance of the wrapper object.

Another way is to move the singleton property to the type that uses it and have it set in the same places where the existing dependency is set.

Most of the time I think I'd prefer the first strategy. In this case I like the second because in this case `GraphRequestConnection` depends on `ApplicationDelegate` but only uses one static method, `isSDKInitialized`. Since we don't care about any other behavior we don't need to know about `ApplicationDelegate` at all. When we set `isSDKInitialized` we can also set a new property to indicate that we can make graph requests.

Additionally this changeset adds some tests around the `overrideGraphAPIVersion` method on `GraphRequestConnection`

Reviewed By: dreamolight

Differential Revision: D26532835

fbshipit-source-id: 98a59a0eedb5d3aad97f4101e1c42b54230a9be1