Backfill GraphRequestConnection Unit Tests 4/n
Summary:
Abstracts the dependency on `ServerConfiguration` since it is only used to obtain an `ErrorConfiguration`. In turn, the `ErrorConfiguration` is only used for obtaining an `ErrorRecoveryConfiguration`. So `GraphRequestConnection` just needs anything that provides something that can be used to get an `ErrorRecoveryConfiguration`.

This diff adds a few new types to achieve this:

* `ErrorRecoveryConfigurationProviding` which describes any type that can provide an `ErrorRecoveryConfiguration` from a given code, subcode, and graph request. We've added default conformance for `FBSDKErrorConfiguration`.

* `FBSDKErrorConfigurationProvider` which serves as the default implementation for the `ErrorConfigurationProviding` protocol. This will either return the current `ErrorConfiguration` from the cached `ServerConfiguration` on `ServerConfigurationManager` or, if
none exists, it will return a new default instance of `ErrorConfiguration`

* `FBSDKDecodableErrorConfiguration` which describes the way an error configuration can be created or updated by encoded remote data.

`GraphRequestConnection` is now initialized by default with an instance of `FBSDKErrorConfigurationProvider`.

This made it easy to test our response parsing and enabled us to uncover some unsafe parsing in how we unpack error responses. This diff adds some safety and tests around that area.

Reviewed By: KylinChang

Differential Revision: D26559834

fbshipit-source-id: afad5656b10a26e319a6a1da3f980023b303fdd8