AuthenticationToken cleanup
Summary:
The main goal of this diff is to remove the the internal dependencies on AuthenticationToken from LoginKit to CoreKit, specifically `claims` and `jti`. LoginKit and CoreKit should be treating each other as third parties libraries and having internal dependencies between them is a bad pattern. (We are already doing that a bunch but it's probably good to not introduce new ones)

For `jti`, we are no longer using it in LoginKit, so I just removed the property. For `claims`, I made it public so LoginKit can use it directly.

Reviewed By: joesus

Differential Revision: D26537072

fbshipit-source-id: f41da429bf67f83f879d345d4e28be651427b19f