Backfill GraphRequestConnection Unit Tests 9/n
Summary:
Currently `GraphRequestConnection` has a dependency on `AppEvents` that's used for logging a single event for an edge case in data parsing.

This adds a couple new types:
* `EventLogging` a protocol which mirrors the `AppEvents` api but uses instance instead of class methods.
* `EventLogger` a concrete type that effectively wraps the `AppEvents` singleton.

This is not the best or final solution to this problem. There's a still a dependency on AppEvents from `EventLogger` but this at least moves the dependency up a level.

We're using `EventLogger` as a default for now but whatever winds up provisioning `GraphRequestConnection`s at the end of this refactor will be able to determine which concrete `EventLogging` is provided.

Also cleans up some unnecessary internal initializer variations.

Reviewed By: jingping2015

Differential Revision: D26645716

fbshipit-source-id: a647c8dc05c9fde226df19ec6f2a2d0492be1544