Backfill GraphRequestConnection Unit Tests 6/n
Summary:
This is intended to abstract the `GraphRequestPiggybackManager` from `GraphRequestConnection`.

To do this we're adding some new types:

* `GraphRequestPiggybackManaging` describes a type that can add piggyback requests to graph request connections.
* `GraphRequestPiggybackManagerProviding` describes a type that can provide a type that conforms to `GraphRequestPiggybackManaging`.
* `GraphRequestPiggybackManagerProvider` - The concrete type to use for `FBSDKGraphRequestConnection`. Provides the class `GraphRequestPiggybackManager` which has default conformance to `GraphRequestPiggybackManaging`.

Haven't found the best pattern for abstracting dependencies with class methods. The right solution here is probably to refactor `GraphRequestPiggybackManager` to be instantiated with the shared state and provide instances of itself as needed. Or maybe this is the right solution, provisioning GraphRequestConnection with a way to request a type that can handle piggybacking.

In any case, this change-set gets us closer to truly inverting the dependency so it can be provided from a central location.

Reviewed By: jingping2015

Differential Revision: D26611208

fbshipit-source-id: 03062fcab847aed74e551dc5f057faedd5c090a3