Backfill GateKeeperManagerTests 2/n
Summary:
In the last diff we added `GraphRequestProtocol` and gave `GraphRequest` default conformance.

This diff prepares for `GateKeeperManager` to use both these types by making `GraphRequestConnecting` and `GraphRequestConnection` deal exclusively in `GraphRequestProtocol`s.

This diff:
* Fills out and makes the `GraphRequestProtocol` public
* Introduces a new protocol `GraphRequestInternalProtocol` to represent the internal methods of `GraphRequest`. Gives `GraphRequest` default conformance in its internal header
* Replaces all method definitions in the project to operate in terms of `GraphRequestProtocol`, where necessary we cast that to `GraphRequestInternalProtocol`. (Since we've added default conformance to both, all the callsites can accept `GraphRequest` still)
* Breaks out some standalone definitions, `GraphRequestHTTPMethod` and `GraphRequestFlags`, into their own headers for organization.

In total, these changes enable a type (in this case `GateKeeperManager`) that has a dependency both on `GraphRequestConnection` and `GraphRequest` to create a `GraphRequestProtocol` and add it to a `GraphRequestConnecting` without it knowing about the concrete types.

Reviewed By: KylinChang

Differential Revision: D26712224

fbshipit-source-id: 2126582bc27a2d606d5584dd834189efdb8bcf3f