Adding Test Specific Macro
Summary:
Introduces a new macro `FBSDKTEST` to our debug configurations.

The goal is to eliminate the possibility that a method used only for testing can be accessed by production code during a DEBUG build since this could lead us to ship a production build that will call an unavailable method.

Reviewed By: jingping2015

Differential Revision: D26610936

fbshipit-source-id: 6483555ca1cf445f34b5ae36d307cb89c54f91e7