Bugfix
Summary:
There's a bug where upgrading from limited to regular login will fail to fetch the profile using the newly available access token.

This is because we check if there's a cached profile with the same user identifier as that on the access token. Since it has the same user identifier and isn't expired, it will not load the new profile.

A profile that is non-expired and matches the currently authenticated user may still be incomplete. This fix formalizes that idea of incompleteness by introducing a new private property `isLimited`. To indicate if the information provided is incomplete in some way. When the property is true, `loadCurrentProfileWithCompletion:` will assume the profile is incomplete and disregard any cached profile. It defaults to false to preserve current caching behavior.

Reviewed By: ppansy

Differential Revision: D26970063

fbshipit-source-id: aa2623ff133b27b6bc0f8d435d849bf3534b3923