Abstracting NSBundle from Settings, AppLinkUtility, InternalUtility
Summary:
We were Swizzling `NSBundle` to be able to test `Settings` and `AppLinkUtility`.

This introduces a new protocol `InfoDictionaryProviding` to stand in for `NSBundle`.

This is a lot of change but it is mostly fixing the API for `FakeBundle` which has been renamed to `TestBundle` and is no longer an `NSBundle` subclass.

Reviewed By: KylinChang

Differential Revision: D27178307

fbshipit-source-id: 66d6d24f7562a598a6b350f57df3fd4a2743901b