Abstracting AccessToken from ApplicationDelegate, Profile
Summary:
AccessToken is all wrong. It is in charge of caching tokens, keeping track of the current token, making network calls to refresh tokens... we can do better.

This introduces two protocols to help us work towards breaking this up. The general plan is to have the existing `AccessToken` conform to two new protocols `AccessTokenProviding` and `AccessTokenSetting`.

I plan to introduce an intermediate type, `TokenWallet`, in the near future that will wrap `AccessToken` and update the semantics so that we can get away from singletons for everything. Overtime we can move concerns related to handling an access token out of `AccessToken` itself and into `TokenWallet`. Ideally in a few releases `AccessToken` can just be a plain old struct.

To Be Determined: `TokenWallet` - type vs protocol? Should there be an `AccessTokenWallet` and an `AuthenticationTokenWallet` or should one known type handle both?

Reviewed By: KylinChang

Differential Revision: D27201309

fbshipit-source-id: 0a755aca196e6a85bce2af2de359b6aa144d3be6