Abstract NSProcessInfo from FBSDKGraphRequestConnection
Summary:
Splits the reliance on NSProcessInfo into two pieces, an operating system version comparer and a mac catalyst determinator.
It seems like overkill but we need to be able to provide these dependencies in a way that isn't arbitrarily coupled.

The use case I'm envisioning here is that user agent can go in its own class and the request connection shouldn't still need to know about the protocol method `isMacCatalystApp`.

Reviewed By: Oliverccccct

Differential Revision: D27240705

fbshipit-source-id: 5ee1f318f745883fcfe65c2ff73e0039a4373d7d