Start moving away from Settings type-based interface
Summary:
Introduces a transitional singleton as a way to change the usage semantics from a type-based interface to an instance-based interface.

Once that is complete then types that use `+[FBSDKSettings foo]` can take an injectable instance of a `FBSDKSettings` until they no longer directly reference a settings type of any kind and instead refer to an injectable dependency for their actual use cases.

The plan is to go from:

ClassWithoutUnderlyingInstance -> ClassRelyingOnUnderlyingInstance -> Instance

Reviewed By: KylinChang

Differential Revision: D27232768

fbshipit-source-id: 21b587a1aeffa68e07a4a554ee85deb8336b548a