add cached ASIdentifierManager to avoid dynamic loading on every didBecomeActive
Summary: Applying in memory cache to the ASIdentifierManager so it only has to be dynamically loaded once on startup and not every time the app is foregrounded. This does NOT cache the advertiser ID itself, which will still be fetched from the ASIdentifierManager on startup and every foregrounding. If a user changes their ad tracking settings while the app is backgrounded, this change will be reflecting as soon as they foreground the app and the ID is fetched from the manager. As per Apple's docs, If the user has turned off ad tracking, an ID with all zeros (00000000-0000-0000-0000-000000000000) will be returned from ASIdentifierManager.

Reviewed By: joesus

Differential Revision: D26887961

fbshipit-source-id: 5030aea3737129afc6d82862fdfa861251ae2273