Bugfix - Cocoapods with generate_multiple_pod_projects
Summary:
We moved auth token claims to the public headers of corekit but didn't remove it from the internal umbrella header. We also didn't add it to the public header for BUCK builds. This fixes both of those oversights.

There was also a weird import of the CoreKit+Internal header in the CoreKit+Internal header. This made no sense so removing.

Differential Revision: D27620983

fbshipit-source-id: 5d97849a4ba0329342ddf2bb5a4c12fb07b5b712