Bugfix - Cocoapods with generate_multiple_pod_projects (#1709)
Summary:
Pull Request resolved: https://github.com/facebook/facebook-ios-sdk/pull/1709

We moved auth token claims to the public headers of corekit but didn't remove it from the internal umbrella header. We also didn't add it to the public header for BUCK builds. This fixes both of those oversights.

There was also a weird import of the CoreKit+Internal header in the CoreKit+Internal header. This made no sense so removing.

Reviewed By: jamestouri

Differential Revision: D27620983

fbshipit-source-id: 24cc10a5bc53568ce9e5cc273011bf147729c441