Rename String coercion utility method
Summary:
This method should probably not exist and definitely shouldn't be used the way it is in most of these cases. This rename intends to surface problematic call sites and avoid future misuse.

Also introduces a new method:
`stringValueOrNil` to be used as a replacement in instances where we just want to ensure that a type returned from the server is a string before calling string  methods on it.

Reviewed By: Oliverccccct

Differential Revision: D27696828

fbshipit-source-id: fec674fa956e79e9cf0f276bba90db96d1660b3e