Invert semantics for AppEventsUtility's advertiserID method
Summary:
This is part of testing and refactoring `CodelessIndexer`. Since it has a dependency on `AppEventsUtility` we're at a decision point where we can decide whether the injected protocol should be instance or type-based.

In general life is a lot easier if we can rely on instances because they are easily reset or modified for tests and do not require cleanup.

This changeset introduces a transitional singleton `shared` to `AppEventsUtility` and updates the call site to use it.

Steps in this process:

[x] Invert semantics of `AppEventsUtility` // this diff
[] Inject instance of `AppEventsUtility` into codeless indexer
[] Replace call sites to use the injected version and add unit tests for those methods

There is also some test cleanup for duplicate mocks:

`stubAdvertiserIdentifierWithIdentifierString` was a duplicate of `stubAppEventsUtilityAdvertiserIDWith`

Reviewed By: Oliverccccct

Differential Revision: D27741028

fbshipit-source-id: 29551b02c88ba147dcd87064d1ddef050be1a170