Stop shadowing dataTask method on NSURLSession
Summary:
This was causing issues in calling the `dataTask(with: URL)` method on instances of `NSURLSession`.

Moves the default conformance to the places that use it so it's not in the header for Basics which is public when using CocoaPods.

Should resolve: https://github.com/facebook/facebook-ios-sdk/issues/1643 and https://github.com/facebook/facebook-ios-sdk/issues/1718

Differential Revision: D27920718

fbshipit-source-id: 58f88e9489ded6db001b5c4f179538025e604236