Stop shadowing dataTask method on NSURLSession (#1721)
Summary:
Pull Request resolved: https://github.com/facebook/facebook-ios-sdk/pull/1721

This was causing issues in calling the `dataTask(with: URL)` method on instances of `NSURLSession`.

Moves the default conformance to the places that use it so it's not in the header for Basics which is public when using CocoaPods.

Should resolve: https://github.com/facebook/facebook-ios-sdk/issues/1643 and https://github.com/facebook/facebook-ios-sdk/issues/1718

Reviewed By: linmx0130

Differential Revision: D27920718

fbshipit-source-id: fdc85dc2f9340285074214d7ea0b2ef9944337ce