Swift Rewrite Pre-work
Summary:
Reorganizes to set us up for a Swift rewrite.

One thing to note about this rewrite is that we cannot simply replace arbitrary files. The dependency graph is extremely important because Swift Package Manager does not support mixed Swift and Objective-C source files.

The high level goal is to maintain backwards compatibility for users importing FBSDKCoreKit while still moving the implementation to Swift under the hood.

To do this, we will need to introduce a new module, LegacyCoreKit which will serve as the source of existing ObjC files.

The existing Swift Package Manager module, FacebookCore, will depend on LegacyCoreKit and will host the top-level Swift files. Top-level in this case refers to those files with the fewest callers from inside the SDK itself; a prime example of this is ApplicationDelegate.

To maintain backwards compatibility for the interfaces, the existing module FBSDKCoreKit will be repurposed to provide wrapper interfaces for FacebookCore.

At the modular level, the dependency graph will look like this:

LegacyCoreKit    the original ObjC implementation undergoing the rewrite
 ↓
FacebookCore    the rewritten native Swift files
 ↓
FBSDKCoreKit    backwards compatible ObjC interfaces

There's still work to do for this including:
* Update CocoaPods to use follow the same structure of relying on an underlying module for the original ObjC implementation
* Update the Xcode targets to follow the same structure and use an underlying module for the original ObjC implementation
* Remove OCMock so that we are forced to restructure the dependencies to be injectable

Reviewed By: Oliverccccct

Differential Revision: D27855791

fbshipit-source-id: f3b5fbbe67492d3a131d475688e91ee8ab51ae6b