Move towards instance-based Logger
Summary:
Our Logger was written to be easy to use with untested Objective-C code. It formats Strings and creates underlying instances to use in its class methods.

This changeset aims to break these patterns by introducing a logging factory so that types will request logger instances from the factory and then use those instances for logging where needed.

Additionally, variadic Objective-C parameters will translate to Swift but it's kind of a pain and there really isn't a compelling reason to include this helper function in any protocol that will be used in defining Swift types. This changes existing ObjC callsites to format their own strings before passing them to the logger since this is how it will work in Swift.

Next steps for cleaning this up:

Get rid of the following class methods:
```
+ (NSUInteger)generateSerialNumber;
+ (void)singleShotLogEntry:(FBSDKLoggingBehavior)loggingBehavior
                  logEntry:(NSString *)logEntry;
+ (void)singleShotLogEntry:(FBSDKLoggingBehavior)loggingBehavior
              timestampTag:(NSObject *)timestampTag
              formatString:(NSString *)formatString, ... NS_FORMAT_FUNCTION(3,4);
+ (void)registerCurrentTime:(NSString *)loggingBehavior
                    withTag:(NSObject *)timestampTag;
+ (void)registerStringToReplace:(NSString *)replace
                    replaceWith:(NSString *)replaceWith;
```
Get rid of the following instance methods since they are only used in GraphRequestConnection. There should be a more specific type to use here instead of forcing all loggers to have / expose these:
```
- (void)appendString:(NSString *)string;
- (void)appendFormat:(NSString *)formatString, ... NS_FORMAT_FUNCTION(1,2);
- (void)appendKey:(NSString *)key value:(NSString *)value;
- (void)emitToNSLog;
```
Inject settings so that Logger does not rely on global state.

Reviewed By: jamestouri

Differential Revision: D28108603

fbshipit-source-id: feb50591fe8596c6fa26850b080d10d46c608200